feat: readme

See merge request s11-webmobile2-sub2/S11P12A508!232

+++ wms_backend/src/main/java/com/a508/wms/productdetail/dto/ProductDetailRequestDto.java
@@ -16,8 +16,8 @@ public class ProductDetailRequestDto {
     private int sellingPrice;
     private Long barcode;
     private ProductStorageTypeEnum productStorageType;
-    private Long size;
-    private Long unit;
+    private int size;
+    private int unit;
     private String name;
     private Long businessId;
 }
